Enhance reflect URL filter implementation
