+++ superset-frontend/src/setup/setupPluginsExtra.ts
@@ -16,11 +16,14 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-import { getExtensionsRegistry } from "@superset-ui/core";
-import KatalonSliceHeaderControls from "src/katalon/KatalonSliceHeaderControls";
+import { getExtensionsRegistry } from '@superset-ui/core';
+import KatalonSliceHeaderControls from 'src/katalon/KatalonSliceHeaderControls';
+import KatalonURLFilters from '../katalon/KatalonURLFilters';
 
 // For individual deployments to add custom overrides
 export default function setupPluginsExtra() {
   const extensionRegistry = getExtensionsRegistry();
+
+  extensionRegistry.set('root.context.provider', KatalonURLFilters);
   extensionRegistry.set('dashboard.slice.header', KatalonSliceHeaderControls);
 }
